requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
requests==2.18.1
scapy-python3==0.21
urllib3==1.21.1
